requirements.txt
tweepy
requests
python-dotenv
psycopg2-binary
